Update memory optimization for 2022
Instead of all the shenanigans with batch/powershell files, it's very simple to use Steam's launch options to set which Java should be used. Also, (now?) Equilinox uses `org.ninjacave.jarsplice.JarSpliceLauncher`, which runs a secondary Java process after the initial one. Thus it is necessary to adjust the `MANIFEST.MF` within the Equilinox jar if you want to increase/change the amount of memory the game is allowed to use.